requirements.txt
requests
python-dotenv
folium
streamlit
streamlit-folium
ultralytics
Pillow
geopandas
shapely
numpy
huggingface_hub
torch
opencv-python
sahi